+++ util-slf4j-api/src/test/scala/com/twitter/util/logging/JavaItem.java
@@ -1,7 +1,7 @@
 package com.twitter.util.logging;
 
 class JavaItem extends SecondItem {
-    private static final Logger LOG = Logger.apply(JavaItem.class);
+    private static final Logger LOG = Logger.getLogger(JavaItem.class);
 
     public JavaItem(String name, String description, int size, int count) {
         super(name, description, size, count);