+++ util-slf4j-api/src/test/scala/com/twitter/util/logging/TestJavaClass.java
@@ -1,7 +1,7 @@
 package com.twitter.util.logging;
 
 class TestJavaClass {
-    private static final Logger LOG = Logger.apply(TestJavaClass.class);
+    private static final Logger LOG = Logger.getLogger(TestJavaClass.class);
 
     public TestJavaClass() {
         LOG.info("Creating new TestJavaClass instance.");