+++ util-slf4j-api/src/test/scala/com/twitter/util/logging/TraitWithLogging.scala
@@ -1,0 +1,11 @@
+package com.twitter.util.logging
+
+abstract class AbstractTraitWithLogging extends TraitWithLogging
+
+trait TraitWithLogging extends Logging {
+
+  def myMethod1: String = {
+    info("In myMethod1")
+    "Hello, World"
+  }
+}